+++ src/main/java/kitEditor/Sound.java
@@ -2,6 +2,8 @@
 
 package kitEditor;
 
+import com.laszlosystems.libresample4j.Resampler;
+
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -12,7 +14,7 @@ public class Sound {
     private static final ArrayList<Clip> clipPool = new ArrayList<>();
     private static final int PLAYBACK_RATE = 48000;
 
-    private static byte[] unpackNibbles(byte[] gbSample) {
+    private static short[] unpackNibbles(byte[] gbSample) {
         byte[] waveData = new byte[gbSample.length * 2];
         int src = 0;
         int dst = 0;
@@ -22,7 +24,16 @@ public class Sound {
             waveData[dst++] = (byte) (0xf0 & sample);
             waveData[dst++] = (byte) ((0x0f & sample) << 4);
         }
-        return waveData;
+
+        short[] s = new short[waveData.length];
+        for (int i = 0; i < s.length; ++i) {
+            int v = waveData[i] & 0xf0;
+            v -= 0x78;
+            v *= Short.MAX_VALUE;
+            v /= 0x78;
+            s[i] = (short)v;
+        }
+        return s;
     }
 
     private static Clip getClip() throws LineUnavailableException {
@@ -37,14 +48,25 @@ public class Sound {
         return newClip;
     }
 
+
+
     static void play(byte[] gbSample, boolean halfSpeed) throws LineUnavailableException, IOException {
         final int sampleRate = halfSpeed ? 5734 : 11468;
-        byte[] waveData = to16Bit(resampleToPlaybackRate(sampleRate, unpackNibbles(gbSample)));
+        byte[] b = toByteArray(resample(sampleRate, PLAYBACK_RATE, unpackNibbles(gbSample)));
         Clip clip = getClip();
-        clip.open(new AudioInputStream(new ByteArrayInputStream(waveData),
+        clip.open(new AudioInputStream(new ByteArrayInputStream(b),
                 new AudioFormat(PLAYBACK_RATE, 16, 1, true, false),
-                waveData.length));
+                b.length / 2));
         clip.start();
+    }
+
+    private static byte[] toByteArray(short[] waveData) {
+        byte[] b = new byte[waveData.length * 2];
+        for (int i = 0; i < waveData.length; ++i) {
+            b[i * 2] = (byte)(waveData[i] & 0xff);
+            b[i * 2 + 1] = (byte)(waveData[i] >> 8);
+        }
+        return b;
     }
 
     static void stopAll() {
@@ -53,36 +75,23 @@ public class Sound {
         }
     }
 
-    private static byte[] to16Bit(float[] waveData) {
-        byte[] dst = new byte[waveData.length * 2];
-        for (int i = 0; i < waveData.length; ++i) {
-            float sf = waveData[i];
-            assert sf >= -1;
-            assert sf <= 1;
-            short si = (short)(sf * Short.MAX_VALUE);
-            dst[i * 2] = (byte)si;
-            dst[i * 2 + 1] = (byte)(si >> 8);
+    public static short[] resample(float inSampleRate, float outSampleRate, short[] samples) {
+        float[] inBuf = new float[samples.length];
+        for (int i = 0; i < inBuf.length; ++i) {
+            inBuf[i] = (float) samples[i] / -Short.MIN_VALUE;
         }
-        return dst;
+
+        double factor = outSampleRate / inSampleRate;
+        float[] outBuf = new float[(int)(inBuf.length * factor + 1)];
+        Resampler resampler = new Resampler(true, factor, factor);
+        Resampler.Result result = resampler.process(factor, inBuf, 0, inBuf.length, true, outBuf, 0, outBuf.length);
+
+        short[] finalBuf = new short[result.outputSamplesGenerated];
+        for (int i = 0; i < finalBuf.length; ++i) {
+            outBuf[i] = Math.max(-1, Math.min(1, outBuf[i]));
+            finalBuf[i] = (short)(outBuf[i] * Short.MAX_VALUE);
+        }
+        return finalBuf;
     }
 
-    // Bespoke resampling routine is a sad necessity since not all systems support
-    // playback of 11468 Hz sample data.
-    private static float[] resampleToPlaybackRate(int srcSampleRate, byte[] srcBuffer) {
-        float[] dstBuffer = new float[(srcBuffer.length * PLAYBACK_RATE) / srcSampleRate];
-        for (int dst = 0; dst < dstBuffer.length; ++dst) {
-            double srcPos = (double)dst * srcSampleRate / PLAYBACK_RATE;
-            int pos1 = (int)Math.floor(srcPos);
-            int pos2 = 1 + pos1;
-            float value = ((int)srcBuffer[pos1]) & 0xf0;
-            if (pos2 < srcBuffer.length) {
-                value *= (1 - srcPos % 1);
-                value += (((int)srcBuffer[pos2]) & 0xf0) * (srcPos % 1);
-            }
-            dstBuffer[dst] = (value - 0x78) / 0x78;
-            assert dstBuffer[dst] <= 1;
-            assert dstBuffer[dst] >= -1;
-        }
-        return dstBuffer;
-    }
 }